requirements.txt
Flask==2.0.3
requests==2.26.0
Pillow==8.4.0
transformers==4.12.2
ultralyticsplus==